app/requirements.txt
Django==3.0.3
gunicorn==20.0.4
pytz==2019.3
djangorestframework==3.11.0
git+https://github.com/hallrizon-io/django3-rest-swagger.git
djangorestframework-jwt-1.11.0